requirements.txt
yt-dlp
youtube-transcript-api
fastapi
uvicorn
python-dotenv
requests
qrcode
pillow
uvicorn
